requirements.txt
# CNN-LSTM Crypto Trading System — Dependencies
# Core
numpy==1.26.4
pandas==2.2.2
pyarrow>=12.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# ML / Deep Learning
torch>=2.0.0
onnx>=1.14.0
onnxruntime>=1.15.0
scikit-learn>=1.3.0

# Feature engineering
numba>=0.58.0
scipy==1.14.0
ta>=0.10.0

# Data fetching
ccxt>=4.0.0
requests>=2.31.0

# Feature selection
shap>=0.42.0
xgboost>=1.7.0
catboost>=1.2.0

# Hyperparameter tuning
optuna>=3.3.0

# Calibration
# (uses sklearn.isotonic — no extra dep)

# Testing
pytest>=7.4.0

# Alerting
python-telegram-bot>=20.0
